+++ ldap/servers/plugins/replication/repl_extop.c
@@ -697,7 +697,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 		connext->repl_protocol_version = REPL_PROTOCOL_50_INCREMENTAL;
 		slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 			"conn=%" NSPRIu64 " op=%d repl=\"%s\": Begin incremental protocol\n",
-			(long long unsigned int)connid, opid, repl_root);
+			connid, opid, repl_root);
 		isInc = PR_TRUE;
 	}
 	else if (strcmp(protocol_oid, REPL_NSDS50_TOTAL_PROTOCOL_OID) == 0)
@@ -723,7 +723,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 		}
 		slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 				"conn=%" NSPRIu64 " op=%d repl=\"%s\": Begin total protocol\n",
-				(long long unsigned int)connid, opid, repl_root);
+				connid, opid, repl_root);
 		isInc = PR_FALSE;
 	}
 	else if (strcmp(protocol_oid, REPL_NSDS71_INCREMENTAL_PROTOCOL_OID) == 0)
@@ -732,7 +732,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 		connext->repl_protocol_version = REPL_PROTOCOL_50_INCREMENTAL;
 		slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 			"conn=%" NSPRIu64 " op=%d repl=\"%s\": Begin 7.1 incremental protocol\n",
-			(long long unsigned int)connid, opid, repl_root);
+			connid, opid, repl_root);
 		isInc = PR_TRUE;
 	}
 	else if (strcmp(protocol_oid, REPL_NSDS71_TOTAL_PROTOCOL_OID) == 0)
@@ -744,7 +744,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 		}
 		slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 				"conn=%" NSPRIu64 " op=%d repl=\"%s\": Begin 7.1 total protocol\n",
-				(long long unsigned int)connid, opid, repl_root);
+				connid, opid, repl_root);
 		isInc = PR_FALSE;
 	}
 	else
@@ -767,7 +767,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
         slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 				"conn=%" NSPRIu64 " op=%d replica=\"%s\": "
 				"Replica is being configured: try again later\n",
-				(long long unsigned int)connid, opid, repl_root);
+				connid, opid, repl_root);
 		response = NSDS50_REPL_REPLICA_BUSY;
 		goto send_response;
 	}
@@ -838,7 +838,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 					slapi_log_error(SLAPI_LOG_FATAL, repl_plugin_name,
 									"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 									"Excessive clock skew from supplier RUV\n",
-									(long long unsigned int)connid, opid, repl_root);
+									connid, opid, repl_root);
 					response = NSDS50_REPL_EXCESSIVE_CLOCK_SKEW;
 					goto send_response;
 				}
@@ -875,7 +875,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
         slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 				"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 				"Replica has same replicaID %d as supplier\n",
-				(long long unsigned int)connid, opid, repl_root, replica_get_rid(replica));
+				connid, opid, repl_root, replica_get_rid(replica));
 		response = NSDS50_REPL_REPLICAID_ERROR;
 		goto send_response;
 	}
@@ -887,7 +887,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 	 */
 	/* junkrc = ruv_get_first_id_and_purl(supplier_ruv, &junkrid, &locking_purl); */
 	PR_snprintf(locking_session, sizeof(locking_session), "conn=%" NSPRIu64 " id=%d",
-			(long long unsigned int)connid, opid);
+			connid, opid);
 	locking_purl = &locking_session[0];
 	if (replica_get_exclusive_access(replica, &isInc, connid, opid,
 									 locking_purl,
@@ -914,7 +914,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 		slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 						"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 						"374 - Starting sleep: connext->repl_protocol_version == %d\n",
-						(long long unsigned int)connid, opid, repl_root, connext->repl_protocol_version);
+						connid, opid, repl_root, connext->repl_protocol_version);
         
 		while (REPL_PROTOCOL_50_INCREMENTAL == connext->repl_protocol_version && i++ < max) {
 			usleep(200000);
@@ -923,7 +923,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 		slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 						"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 						"374 - Finished sleep: connext->repl_protocol_version == %d\n",
-						(long long unsigned int)connid, opid, repl_root, connext->repl_protocol_version);
+						connid, opid, repl_root, connext->repl_protocol_version);
 	}
 #endif
 
@@ -1018,7 +1018,7 @@ multimaster_extop_StartNSDS50ReplicationRequest(Slapi_PBlock *pb)
 				"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 				"Unexpected update protocol received: %d.  "
 				"Expected incremental or total.\n",
-				(long long unsigned int)connid, opid, repl_root, connext->repl_protocol_version);
+				connid, opid, repl_root, connext->repl_protocol_version);
 		goto send_response;
     }
 
@@ -1057,7 +1057,7 @@ send_response:
 		slapi_log_error (resp_log_level, repl_plugin_name,
 			"conn=%" NSPRIu64 " op=%d replica=\"%s\": "
 			"Unable to acquire replica: error: %s%s\n",
-			(long long unsigned int)connid, opid,
+			connid, opid,
 			(replica ? slapi_sdn_get_dn(replica_get_root(replica)) : "unknown"),
 			protocol_response2string (response), purlstr);
 
@@ -1109,7 +1109,7 @@ send_response:
 	slapi_log_error(SLAPI_LOG_REPL, repl_plugin_name,
 					"conn=%" NSPRIu64 " op=%d repl=\"%s\": "
 					"%s: response=%d rc=%d\n",
-					(long long unsigned int)connid, opid, repl_root,
+					connid, opid, repl_root,
 					is90 ? "StartNSDS90ReplicationRequest" :
 					"StartNSDS50ReplicationRequest", response, rc);
 	slapi_send_ldap_result(pb, LDAP_SUCCESS, NULL, NULL, 0, NULL);
