+++ ldap/servers/slapd/result.c
@@ -2002,7 +2002,7 @@ log_result( Slapi_PBlock *pb, Operation *op, int err, ber_tag_t tag, int nentrie
 							  "conn=%" NSPRIu64 " op=%d RESULT err=%d"
 							  " tag=%" BERTAG_T " nentries=%d etime=%s%s%s"
 							  ", SASL bind in progress\n",
-							  (long long unsigned int)op->o_connid,
+							  op->o_connid,
 							  op->o_opid,
 							  err, tag, nentries,
 							  etime,
@@ -2034,7 +2034,7 @@ log_result( Slapi_PBlock *pb, Operation *op, int err, ber_tag_t tag, int nentrie
 							  "conn=%" NSPRIu64 " op=%d RESULT err=%d"
 							  " tag=%" BERTAG_T " nentries=%d etime=%s%s%s"
 							  " dn=\"%s\"\n",
-							  (long long unsigned int)op->o_connid,
+							  op->o_connid,
 							  op->o_opid,
 							  err, tag, nentries,
 							  etime,
@@ -2059,7 +2059,7 @@ log_result( Slapi_PBlock *pb, Operation *op, int err, ber_tag_t tag, int nentrie
 			slapi_log_access( LDAP_DEBUG_STATS,
 							  "conn=%" NSPRIu64 " op=%d RESULT err=%d"
 							  " tag=%" BERTAG_T " nentries=%d etime=%s%s%s\n",
-							  (long long unsigned int)op->o_connid,
+							  op->o_connid,
 							  op->o_opid,
 							  err, tag, nentries,
 							  etime,
@@ -2123,7 +2123,7 @@ log_entry( Operation *op, Slapi_Entry *e )
 	if ( !internal_op )
 	{
 		slapi_log_access( LDAP_DEBUG_STATS2, "conn=%" NSPRIu64 " op=%d ENTRY dn=\"%s\"\n",
-			(long long unsigned int)op->o_connid, op->o_opid,
+			op->o_connid, op->o_opid,
 			slapi_entry_get_dn_const(e));
 	}
 	else
@@ -2148,7 +2148,7 @@ log_referral( Operation *op )
 	if ( !internal_op )
 	{
 		slapi_log_access( LDAP_DEBUG_STATS2, "conn=%" NSPRIu64 " op=%d REFERRAL\n",
-			(long long unsigned int)op->o_connid, op->o_opid );
+			op->o_connid, op->o_opid );
 	}
 	else
 	{
